requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.4.0
automat==25.4.16
beautifulsoup4==4.14.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.4.0
defusedxml==0.7.1
distro==1.9.0
filelock==3.25.0
flatbuffers==25.12.19
google-auth==2.49.0
google-genai==1.66.0
greenlet==3.3.2
groq==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
jmespath==1.1.0
lxml==6.0.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.4
onnxruntime==1.24.3
packaging==26.0
pandas==2.2.3
parsel==1.11.0
pdfminer-six==20251230
pdfplumber==0.11.9
pillow==12.1.1
protego==0.6.0
protobuf==7.34.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.13.1
pydispatcher==2.0.7
pymupdf==1.27.1
pymupdf-layout==1.27.1
pymupdf4llm==0.3.4
pyopenssl==25.1.0
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.3
queuelib==1.9.0
requests==2.32.3
requests-file==3.0.1
rsa==4.9.1
scrapy==2.14.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.48
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
tldextract==5.3.1
twisted==25.5.0
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
w3lib==2.4.0
websockets==16.0
zope-interface==8.2
